requirements.txt
langchain==0.0.350
langchain-core==0.1.23
langchain-community==0.0.20
langchain-text-splitters==0.0.1

langchain-groq==0.0.1
langchain-google-genai==1.0.1

pydantic==1.10.13

faiss-cpu==1.7.4
streamlit==1.32.2
python-dotenv==1.0.1